Nisan19/flaskapp/requirements.txt
flask
gunicorn
Flask-PyMongo==2.2.0